+++ packages/connectkit/src/defaultConfig.ts
@@ -1,54 +1,12 @@
-import { type CreateConnectorFn, fallback, http, webSocket } from 'wagmi';
+import { type CreateConnectorFn } from 'wagmi';
 import { type CreateConfigParameters } from '@wagmi/core';
-import { type Chain, mainnet, polygon, optimism, arbitrum } from 'wagmi/chains';
-import { type HttpTransport, type WebSocketTransport } from 'viem';
+import { mainnet, polygon, optimism, arbitrum } from 'wagmi/chains';
 
 import defaultConnectors from './defaultConnectors';
-import { chainConfigs } from './constants/chainConfigs';
 
-const rpcUrls: {
-  [key: string]: {
-    [chainId: string]: string;
-  };
-} = {
-  infura: {
-    1: 'https://mainnet.infura.io/v3/',
-    5: 'https://goerli.infura.io/v3/',
-    11155111: 'https://sepolia.infura.io/v3/',
-  },
-};
-
-const getTransport = ({
-  chain,
-  provider = 'public',
-  apiKey,
-}: {
-  chain: Chain;
-  provider: 'alchemy' | 'infura' | 'public';
-  apiKey: string;
-}): HttpTransport | WebSocketTransport => {
-  const supportedChain = chainConfigs.find((c) => c.id === chain.id);
-  if (supportedChain?.rpcUrls) {
-    if (provider === 'alchemy') {
-      if (supportedChain.rpcUrls?.alchemy?.http) {
-        return http(supportedChain.rpcUrls?.alchemy?.http + apiKey);
-      } else {
-        return webSocket(supportedChain.rpcUrls?.alchemy?.webSocket + apiKey);
-      }
-    } else if (provider === 'infura') {
-      if (supportedChain.rpcUrls?.infura?.http) {
-        return http(supportedChain.rpcUrls?.infura?.http + apiKey);
-      } else {
-        return webSocket(supportedChain.rpcUrls?.infura?.webSocket + apiKey);
-      }
-    }
-  }
-  return http();
-};
-
+// TODO: Move these to a provider rather than global variable
 let globalAppName: string;
 let globalAppIcon: string;
-
 export const getAppName = () => globalAppName;
 export const getAppIcon = () => globalAppIcon;
 
@@ -57,16 +15,10 @@ type DefaultConfigProps = {
   appIcon?: string;
   appDescription?: string;
   appUrl?: string;
-  chains?: CreateConfigParameters['chains'];
-  transports?: CreateConfigParameters['transports'];
-  connectors?: CreateConnectorFn[];
 
-  // API keys
-  alchemyId?: string;
-  infuraId?: string;
   // WC 2.0 requires a project ID (get one here: https://cloud.walletconnect.com/sign-in)
   walletConnectProjectId: string;
-};
+} & CreateConfigParameters;
 
 const defaultConfig = ({
   appName = 'ConnectKit',
@@ -76,29 +28,16 @@ const defaultConfig = ({
   chains = [mainnet, polygon, optimism, arbitrum],
   connectors,
   walletConnectProjectId,
-  alchemyId,
-  infuraId,
+  ...props
 }: DefaultConfigProps): CreateConfigParameters => {
   globalAppName = appName;
   if (appIcon) globalAppIcon = appIcon;
 
-  const transports: CreateConfigParameters['transports'] = {};
-  Object.keys(chains).forEach((key, index) => {
-    const chain = chains[index];
-    const urls: (HttpTransport | WebSocketTransport)[] = [];
-    if (alchemyId)
-      urls.push(
-        getTransport({ chain, provider: 'alchemy', apiKey: alchemyId })
-      );
-    if (infuraId)
-      urls.push(getTransport({ chain, provider: 'infura', apiKey: infuraId }));
-    urls.push(http());
-    transports[chain.id] = fallback(urls);
-  });
+  // TODO: nice to have, automate transports based on chains
 
   const config: CreateConfigParameters = {
+    ...props,
     chains,
-    transports,
     connectors:
       connectors ??
       defaultConnectors({
